requirements.txt
chromadb==0.6.3
langchain==0.3.18
langchain_chroma==0.2.1
langchain_community==0.3.17
langchain_core==0.3.34
langchain_ollama==0.2.3
pandas==2.2.3
python_pptx==1.0.2
streamlit==1.42.0
